app/pyproject.toml
[tool.poetry]
name = "base-django"
version = "0.1.0"
description = ""
authors = ["Cesar Freire <cesar.freire@magazord.com.br>"]
readme = "README.md"
packages = [{include = "base_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
django-allauth = "^0.58.1"
crispy-bootstrap5 = "^2023.10"
environs = "^9.5.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
